server/src/tests/molecule/requirements.txt
molecule
molecule-plugins
ansible
pytest
pytest-testinfra
